chore: automate profile screen for android and ios
update previous profile test screen name to settings screen
