+++ tests/android/pages/android_profile.py
@@ -363,3 +363,19 @@ class AndroidProfile(AndroidBasePage):
         return self.global_contents.get_element_by_text(
             self.driver,
             android_elements.profile_dates_calendar_label)
+
+    def get_edit_profile_button(self):
+        """
+        Returns:
+            element: edit profile button element
+        """
+
+        self.global_contents.wait_for_element_visibility(
+            self.driver,
+            android_elements.edit_profile_button
+        )
+
+        return self.global_contents.wait_and_get_element(
+            self.driver,
+            android_elements.edit_profile_button
+        )
